Update init to configure connection_pool properly
Currently a blank connection pool is being initialized when a new
MRedis class is instantiated, which causes issues when your host
or port is different from the default, because the connection itself
is using only defaults.

This change allows the Redis class to create a new connection_pool
with the correct parameters for each server requested.